Separate property bag tests.
